requirements.txt
tweepy==4.14.0
python-dotenv==1.0.0
openai==0.28.1
schedule==1.2.1
python-dateutil==2.8.2
requests==2.31.0
